[PORT] delivery_carrier_label_postlogistics port to 9.0 - Reactivate - Move files in views and models - Adapt XML views - Update README.rst
[PORT] delivery_carrier_label_postlogistics - move license config in a separate view list. This fix an issue of write on related One2many in res.config.settings object

Update README.rts with links and requirements

Add missing set default on licenses

Use short license header

Set module category

<openerp> replaced by <odoo> tag in xml files

Update README.rst

Fix postlogistics tests
