+++ delivery_carrier_label_postlogistics/__openerp__.py
@@ -1,40 +1,24 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Yannick Vaucher
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2013-2016 Yannick Vaucher (Camptocamp SA)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {'name': 'PostLogistics Labels WebService',
- 'version': '8.0.1.1.0',
+ 'version': '9.0.1.1.0',
  'author': "Camptocamp,Odoo Community Association (OCA)",
  'maintainer': 'Camptocamp',
  'license': 'AGPL-3',
- 'category': 'version',
+ 'category': 'Delivery',
  'complexity': 'normal',
  'depends': ['base_delivery_carrier_label'],
  'website': 'http://www.camptocamp.com/',
- 'data': ['res_partner_data.xml',
-          'delivery_data.xml',
-          'delivery_view.xml',
-          'res_config_view.xml',
+ 'data': ['data/res_partner.xml',
+          'data/delivery.xml',
+          'views/delivery.xml',
+          'views/postlogistics_license.xml',
+          'views/res_config.xml',
           'security/ir.model.access.csv',
           ],
  'tests': [],
- 'installable': False,
+ 'installable': True,
  'auto_install': False,
  'application': True,
  'external_dependencies': {
