+++ delivery_carrier_label_postlogistics/models/__init__.py
@@ -1,0 +1,5 @@
+from . import company
+from . import res_config
+from . import delivery
+from . import stock
+from . import stock_quant_package
