+++ delivery_carrier_label_postlogistics/models/delivery.py
@@ -1,23 +1,6 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Yannick Vaucher
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2013-2016 Yannick Vaucher (Camptocamp SA)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from lxml import etree
 from openerp import models, fields, api
 
@@ -35,7 +18,8 @@ class PostlogisticsLicense(models.Model):
                          required=True)
     company_id = fields.Many2one(comodel_name='res.company',
                                  string='Company',
-                                 required=True)
+                                 required=True,
+                                 default=lambda self: self.env.user.company_id)
     sequence = fields.Integer(
         string='Sequence',
         help="Gives the sequence on company to define priority on license "
