+++ delivery_carrier_label_postlogistics/postlogistics/web_service.py
@@ -1,30 +1,12 @@
 # -*- coding: utf-8 -*-
-##############################################################################
-#
-#    Author: Yannick Vaucher
-#    Copyright 2013 Camptocamp SA
-#
-#    This program is free software: you can redistribute it and/or modify
-#    it under the terms of the GNU Affero General Public License as
-#    published by the Free Software Foundation, either version 3 of the
-#    License, or (at your option) any later version.
-#
-#    This program is distributed in the hope that it will be useful,
-#    but WITHOUT ANY WARRANTY; without even the implied warranty of
-#    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#    GNU Affero General Public License for more details.
-#
-#    You should have received a copy of the GNU Affero General Public License
-#    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#
-##############################################################################
+# © 2013-2015 Yannick Vaucher (Camptocamp SA)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 import re
 import logging
 from PIL import Image
 from StringIO import StringIO
 
-from openerp.osv import orm
-from openerp.tools.translate import _
+from openerp import exceptions, _
 
 _compile_itemid = re.compile(r'[^0-9A-Za-z+\-_]')
 _logger = logging.getLogger(__name__)
@@ -78,8 +60,7 @@ class PostlogisticsWebService(object):
         except Exception as e:
             # if authentication error
             if isinstance(e[0], tuple) and e[0][0] == 401:
-                raise orm.except_orm(
-                    _('Error 401'),
+                raise exceptions.Warning(
                     _('Authorization Required\n\n'
                       'Please verify postlogistics username and password in:\n'
                       'Configuration -> Postlogistics'))
