+++ delivery_carrier_label_postlogistics/tests/__init__.py
@@ -1,3 +1,1 @@
-# -*- coding: utf-8 -*-
-
 from . import test_postlogistics
