+++ delivery_carrier_label_postlogistics/views/postlogistics_license.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<odoo>
+
+    <record id="postlogistics_license_view_list" model="ir.ui.view">
+      <field name="name">postlogistics.license.view_list</field>
+      <field name="model">postlogistics.license</field>
+      <field name="arch" type="xml">
+        <tree editable="bottom">
+          <field name="name"/>
+          <field name="number"/>
+          <field name="sequence" invisible="True"/>
+          <field name="company_id" groups="base.group_multi_company"/>
+        </tree>
+      </field>
+    </record>
+
+    <record id="action_postlogistics_licenses" model="ir.actions.act_window">
+      <field name="name">Postlogistics Licenses</field>
+      <field name="type">ir.actions.act_window</field>
+      <field name="res_model">postlogistics.license</field>
+      <field name='view_type'>form</field>
+      <field name='view_mode'>tree</field>
+    </record>
+
+    <menuitem id="menu_postlogistics_licenses" name="Postlogistics Licenses"
+              parent="base_delivery_carrier_label.menu_carriers_config"
+              sequence="20" action="action_postlogistics_licenses"/>
+
+</odoo>
